Properly handle slides at the beginning or end positions in a staff.
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@793 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/powertabdocument/staff.cpp
@@ -416,7 +416,7 @@ Position* Staff::GetPositionByPosition(uint32_t voice, uint32_t index) const
 
 /// Finds the position index of the next position
 /// @throw std::out_of_range if the position does not exist in this staff, or if the voice is invalid
-size_t Staff::GetIndexOfNextPosition(uint32_t voice, boost::shared_ptr<const System> system,
+uint32_t Staff::GetIndexOfNextPosition(uint32_t voice, boost::shared_ptr<const System> system,
                                      const Position* position) const
 {
     if (!IsValidVoice(voice))
